pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "building-infra-sims"
version = "0.1.0"
description = "Building infrastructure protocol simulators for BACnet/IP and Modbus TCP testing"
requires-python = ">=3.11"
dependencies = [
    "bacpypes3>=0.0.99",
    "pymodbus>=3.6",
    "httpx>=0.27",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "typer>=0.12",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "BAC0>=2024.0",
    "ruff>=0.4",
]

[project.scripts]
bsim = "building_infra_sims.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
